+++ src/App.tsx
@@ -11,18 +11,18 @@ import { WorkoutLogs } from './components/WorkoutLogs';
 import { Settings } from './components/Settings';
 import { Login } from './components/Auth/Login';
 import { SignUp } from './components/Auth/SignUp';
-import { SplashScreen } from './components/SplashScreen';
+import { SplashScreen } from './components/others/SplashScreen';
 import ResetPasswordPage from './pages/ResetPasswordPage';
 import ResetPasswordConfirmPage from './pages/ResetPasswordConfirmPage';
-import { LogDayLogo } from './components/LogDayLogo';
+import { LogDayLogo } from './components/others/LogDayLogo';
 import { ContactForm } from './components/ContactForm';
-import { UpdateNotification } from './components/UpdateNotification';
+import { UpdateNotification } from './components/others/UpdateNotification';
 import { WorkoutSkeleton } from './components/WorkoutSkeleton';
 import RoutinesPage from './pages/routines';
 import ProfilePage from './pages/ProfilePage';
 import { Capacitor } from '@capacitor/core';
 import { capacitorService } from './services/capacitor';
-import { WhatsNewModal } from './components/WhatsNewModal';
+import { WhatsNewModal } from './components/others/WhatsNewModal';
 import Lottie from 'lottie-react';
 import loaderAnimation from './assets/animations/loader-app.json';
