+++ src/components/ExerciseList.tsx
@@ -5,8 +5,8 @@ import { motion, AnimatePresence } from 'framer-motion';
 import { Exercise } from '../types/exercise';
 import { useWorkout } from '../context/WorkoutContext';
 import { generateWorkoutName } from '../utils/workoutNameGenerator';
-import { OngoingWorkoutMessage } from './OngoingWorkoutMessage';
-import { InstallAppToast } from './InstallAppToast';
+import { OngoingWorkoutMessage } from './others/OngoingWorkoutMessage';
+import { InstallAppToast } from './others/InstallAppToast';
 import { AddExerciseModal } from './AddExerciseModal';
 import { exerciseService } from '../services/exerciseService';
 import { supabaseService } from '../services/supabaseService';
