+++ src/components/others/LogDayLogo.tsx
