+++ src/components/others/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
+import { useAuth } from '../../context/AuthContext';
 
 export default function ProtectedRoute() {
   const { user } = useAuth();
